+++ demos/common/Link.vue
@@ -41,7 +41,7 @@ const isActive = computed(() => {
 	font-weight: 400;
 	padding-left: 16px;
 	border-left: 4px solid transparent;
-	color: #595b66;
+	color: var(--demo-link-fg);
 	list-style: none;
 	cursor: pointer;
 	text-decoration: none;
@@ -54,7 +54,7 @@ const isActive = computed(() => {
 .demo.active,
 .demo:hover {
 	font-weight: 500;
-	color: #42454d;
-	background-color: #f1f1f1;
+	color: var(--demo-link-active-fg);
+	background-color: var(--demo-link-active-bg);
 }
 </style>